feat(orders): add common order templates
simple and complex orders (in the previous commit) are fairly difficult to build, even with the
provided OrderBuilder classes. To make it easier for people who are familiar with common options
strategies and want easy ways to place orders without wanting the full complexities of the
ComplexOrder or SimpleOrder, add in several template functions to pre-build common strategies. These
strategies (like Iron Condor, Vertical Spreads, Butterfly, etc.) are common strategies used by
TastyTrade and a large number of traders. While this isn't a comprehensive list and the
configurability isn't verbose, the design is aimed towards simplicity first.